requirements.txt
selenium>=4.18.0
requests>=2.31.0
PySide6>=6.0.0
qtawesome>=1.2.1
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
pillow>=9.0.0
